[Refactor] tripConverter로 기능 분리
